% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_coordinates.R
\name{convert_coordinates}
\alias{convert_coordinates}
\title{Convert coordinates from Gauss-Krueger to WGS84}
\usage{
convert_coordinates(.Data = NULL, .gk.rw = NULL, .gk.hw = NULL)
}
\description{
Convert coordinates from Gauss-Krueger to WGS84
}
\examples{
gk.rw <- 3557875
gk.hw <- 5541342
convert_coordinates(.gk.rw = gk.rw, .gk.hw = gk.hw)
# lat <- 50.005600; lng <- 9.806272

Data <- data.table(gk.rw = c(3557875, 4454942), gk.hw = c(5541342, 5532305))
convert_coordinates(Data)
}
