% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tk25.R
\name{coordinates_to_tk25}
\alias{coordinates_to_tk25}
\title{Coordinates to TK25}
\usage{
coordinates_to_tk25(.longitude = NULL, .latitude = NULL)
}
\arguments{
\item{.longitude}{A numeric vector of longitudes.}

\item{.latitude}{A numeric vector of latitudes.}
}
\description{
Coordinates to TK25
}
\examples{
coordinates_to_tk25(11, 48)
}
