% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platform.R
\name{fake_platform}
\alias{fake_platform}
\title{make \code{platform} from GPL accession to test \code{\link[=guess_platform_type]{guess_platform_type()}}}
\usage{
fake_platform(accession, gpl_metas)
}
\description{
make \code{platform} from GPL accession to test \code{\link[=guess_platform_type]{guess_platform_type()}}
}
\keyword{internal}
