% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geo_info.R
\name{get_geo_info}
\alias{get_geo_info}
\title{Get geographic information for coordinates.}
\usage{
get_geo_info(
  .longitude,
  .latitude,
  .naturraum = NULL,
  .boundaries = NULL,
  .elevation = NULL,
  .tk25 = TRUE
)
}
\arguments{
\item{.longitude}{A numeric vector of longitudes.}

\item{.latitude}{A numeric vector of latitudes.}

\item{.naturraum}{Optional sf polygon of naturraum data.}

\item{.boundaries}{Optional sf polygon of administrative boundaries.}

\item{.elevation}{Optional raster of elevation values.}

\item{.tk25}{Logical value. Should tk25 info be included?}
}
\value{
A data.table with one row per coordinate pair.
}
\description{
This functions intersects coordinates with a set of shapefiles.
}
\examples{
get_geo_info(
  .longitude = 11, .latitude = 48,
  .naturraum = naturraum_bayern,
  .boundaries = boundaries_bayern
)
}
