% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geographic_data.R
\name{get_geographic_data}
\alias{get_geographic_data}
\title{Get geographic information for coordinates}
\usage{
get_geographic_data(.Data, .countries = list())
}
\arguments{
\item{.Data}{A data.table containing latitude and longitude columns.}

\item{.countries}{A list with geographic data from \code{\link[raster:getData]{raster::getData()}}.}
}
\value{
a list containing named entries NAME_0, NAME_1, NAME_2, NAME_3, NAME_4. See \code{\link[raster:getData]{raster::getData()}}.
}
\description{
Get geographic information for coordinates
}
\examples{
library(data.table)
data <- data.table(latitude = 48, longitude = 11)
geographic_data_de <- raster::getData("GADM", country = "DE", level = 2)
get_geographic_data(data, .countries = list(geographic_data_de))
file.remove("GADM_2.8_DEU_adm2.rds")
}
\seealso{
\code{\link[raster:getData]{raster::getData()}}
}
