% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_naturraum.R
\name{get_naturraum}
\alias{get_naturraum}
\title{Get naturraum (Bavaria)}
\usage{
get_naturraum(.Data = NULL, .latitude = NULL, .longitude = NULL)
}
\arguments{
\item{.Data}{A data.table.
You can either specify an input data.table with columns named \code{longitude}
and \code{latitude} or pass on two equal-length vectors to \code{.longitude} and \code{.latitude}.}

\item{.latitude}{Numeric vector of latitudes.}

\item{.longitude}{Numeric vector of longitudes.}
}
\value{
A data.table with three columns.
}
\description{
Get naturraum (Bavaria)
}
\examples{
get_naturraum(data.table(latitude = 48, longitude = 11))
}
