% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tk25.R
\name{get_tk25}
\alias{get_tk25}
\alias{get_tk25_number}
\alias{get_tk25_quadrant_number}
\alias{get_tk25_coordinates}
\alias{get_tk25_name}
\alias{get_tk25_info}
\title{Get TK25 information.}
\usage{
get_tk25(.Data)

get_tk25_number(.latitude, .longitude)

get_tk25_quadrant_number(.latitude, .longitude, .tk25.center.latitude,
  .tk25.center.longitude)

get_tk25_coordinates(.tk25.number, .quadrant = NULL)

get_tk25_name(.tk25.number)

get_tk25_info(.tk25.number, .quadrant = NULL)
}
\arguments{
\item{.Data}{A data.table with latitude and longitude columns.}

\item{.latitude}{Numeric vector of latitudes.}

\item{.longitude}{Numeric vectors of longitudes.}

\item{.tk25.center.latitude}{Numeric vector of latitudes. Center of TK25.}

\item{.tk25.center.longitude}{Numeric vector of longitudes. Center of TK25.}

\item{.tk25.number}{Character or numeric vector. Number of TK25 square,
e.g. "1648" or "1648_1" (allowed in get_tk25_info).}

\item{.quadrant}{Character, numeric or logical vector. Number of TK25 Quadrant, between 1 and 4.}
}
\description{
TK25 (Topographische Karte) is a german topographic named raster.
Get name, number or coordinates of corners and center points respectively.
}
\examples{
library(data.table)
get_tk25(data.table(latitude = 48, longitude = 11))
get_tk25_number(48, 11)
get_tk25_quadrant_number(53.6, 10.3, 53.6485, 10.24883)
get_tk25_coordinates("2327")
get_tk25_coordinates("2327", "1")
get_tk25_name("2327")
get_tk25_info("2327")
}
