% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tk25.R
\name{get_tk25_full}
\alias{get_tk25_full}
\title{Get TK25 information}
\usage{
get_tk25_full(.Data = NULL, .latitude = NULL, .longitude = NULL)
}
\arguments{
\item{.Data}{A data.table.
You can either specify an input data.table with columns named \code{longitude}
and \code{latitude} or pass on two equal-length vectors to \code{.longitude} and \code{.latitude}.}

\item{.latitude}{Numeric vector of latitudes.}

\item{.longitude}{Numeric vector of longitudes.}
}
\description{
Get TK25 information
}
\examples{
get_tk25_full(.latitude = 47.5, .longitude = 11)
}
