% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftp.R
\name{gpl_soft_ftp}
\alias{gpl_soft_ftp}
\alias{gse_soft_ftp}
\alias{gse_matrix_ftp}
\title{ftp path to GEO raw data file}
\usage{
gpl_soft_ftp(accession)

gse_soft_ftp(accession)

gse_matrix_ftp(accession)
}
\arguments{
\item{accession}{character. like 'GPL6947', 'GSE19161'}
}
\value{
character. ftp path to file
}
\description{
\code{gpl_soft_ftp()} for GPL SOFT file, \code{gse_soft_ftp()} for GSE SOFT file, \code{gse_matrix_ftp()} for GSE matrix file
}
\examples{
gpl_soft_ftp('GPL6947')

gse_soft_ftp('GSE19161')

gse_matrix_ftp('GSE19161')

}
