% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-gse.R
\name{parse_gse_soft}
\alias{parse_gse_soft}
\title{parse GSE SOFT file (contains platform annotation)}
\usage{
parse_gse_soft(soft_file, verbose = T)
}
\arguments{
\item{path}{string. path to the SOFT file}
}
\value{
list
\enumerate{
\item accession: string. platform accession
\item info: tibble. chip column name information
\item table: tibble. chip annotation
}
}
\description{
parse GSE SOFT file (contains platform annotation)
}
\details{
for now, we drop probes which haven't been mapped to a symbool (mapping to multiple symbols is okay).
}
\examples{
parse_gse_soft(system.file('extdata/GSE19161_family.soft.gz', package = 'rGEO'), verbose = F)

parse_gse_soft(system.file('extdata/GSE51280_family.soft.gz', package = 'rGEO'))

}
\seealso{
Other read raw data: 
\code{\link{read_gse_matrix}()},
\code{\link{read_gse_soft}()}
}
\concept{read raw data}
