% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-info.R
\name{read_gpl_html}
\alias{read_gpl_html}
\title{read GPL information html file}
\usage{
read_gpl_html(file)
}
\arguments{
\item{file}{string.}
}
\value{
list
\enumerate{
\item info: tibble (\code{info}, 'name')
\item sample: tibble, part of the platform table
}
}
\section{to do}{
 when info is \code{NULL}, shouldn't read sample table (in this case you will read supplementary file table)
}

\examples{
NULL
}
