% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-gse.R
\name{read_gse_soft}
\alias{read_gse_soft}
\title{read GSE soft file (contains platform annotation)}
\usage{
read_gse_soft(soft_file, verbose = F)
}
\arguments{
\item{path}{string. path to file}
}
\value{
tibble. the first variable is probe id and the second one is hugo gene symbol
}
\details{
for now, we drop probes which haven't been mapped to a symbool (mapping multiple symbols is okay)
}
\examples{

}
\seealso{
\link{write_chip}, \link{parse_gse_soft}
}
