% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-info.R
\name{read_summary}
\alias{read_summary}
\title{read GEO datasets summary txt file as a tibble}
\usage{
read_summary(path)
}
\arguments{
\item{path}{string. path to the summary file}
}
\value{
tibble
\enumerate{
\item accession
\item type
\item species
\item ftp
\item platform
\item data_type
\item title
\item description
}
}
\examples{
read_summary('data-raw/gds_result.txt')

}
