% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-gse.R
\name{rm_problematic_row}
\alias{rm_problematic_row}
\title{remove problematic rows after you \link[readr:read_delim]{readr::read_*()} a file}
\usage{
rm_problematic_row(data)
}
\arguments{
\item{data}{tibble}
}
\value{
tibble
}
\description{
remove problematic rows after you \link[readr:read_delim]{readr::read_*()} a file
}
\examples{
problematic_df <- readr::read_tsv('a\tb\tc\n1\n2\t3\n1\t2\t3')

problematic_df

rm_problematic_row(problematic_df)

}
