% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tk25.R
\docType{data}
\name{tk25}
\alias{tk25}
\title{TK25}
\format{
data.frame with 6 columns
}
\description{
TK25 (Topographische Karte 1:25000) is a german topographic named raster.
Contains number, name and west, south, north, east border of each German TK25.
}
\keyword{data}
