% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tk25.R
\name{tk25_to_coordinates}
\alias{tk25_to_coordinates}
\title{TK25 to coordinates}
\usage{
tk25_to_coordinates(.tk25, .quadrant = FALSE, .viertelquadrant = FALSE)
}
\arguments{
\item{.tk25}{TK25 Number. String in format "7933 33".}

\item{.quadrant}{Boolean value.}

\item{.viertelquadrant}{Boolean value.}
}
\description{
TK25 to coordinates
}
\examples{
tk25_to_coordinates("7932 33 Utting")
}
