% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_coordinates.R
\name{convert_coordinates}
\alias{convert_coordinates}
\title{Convert coordinates from Gauss-Krueger to WGS84.}
\usage{
convert_coordinates(.gk.rw = NULL, .gk.hw = NULL)
}
\arguments{
\item{.gk.rw}{A numeric vector of Gauss Krueger Rechtswert coordinates.}

\item{.gk.hw}{A numeric vector of Gauss Krueger Hochwert coordinates.}
}
\description{
Convert coordinates from Gauss-Krueger to WGS84.
}
\examples{
convert_coordinates(.gk.rw = 3557875, .gk.hw = 5541342)
# lat <- 50.0056; lng <- 9.806289
}
