% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tk25.R
\name{coordinates_to_tk25}
\alias{coordinates_to_tk25}
\title{Coordinates to TK25}
\usage{
coordinates_to_tk25(.longitude = NULL, .latitude = NULL, .Data = NULL)
}
\arguments{
\item{.longitude}{Numeric vector of longitudes.}

\item{.latitude}{Numeric vector of latitudes.}

\item{.Data}{A data.table.
You can either specify an input data.table with columns named \code{longitude}
and \code{latitude} or pass on two equal-length vectors to \code{.longitude} and \code{.latitude}.}
}
\description{
Coordinates to TK25
}
\examples{
coordinates_to_tk25(11, 48)
}
