% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwn_shp_zip.R
\name{dwn_shp_zip}
\alias{dwn_shp_zip}
\title{dwn_shp_zip}
\usage{
dwn_shp_zip(
  url,
  cname_old = NA,
  cname_new = NA,
  crs = NULL,
  out_path = NULL,
  top = NULL,
  verbose = FALSE
)
}
\arguments{
\item{url}{a url for a zipped file containing one \emph{shapefile}}

\item{cname_old}{When not \code{NA}, and a so-called column exists, the returned object will have only that column}

\item{cname_new}{When not \code{NA},}

\item{crs}{If not \code{NULL},}

\item{out_path}{If not \code{NULL},}

\item{top}{If not \code{NULL}, returns only the first \code{top} complete records, disregarding all other options.}

\item{verbose}{When \code{TRUE} highlight with messages all operations.}
}
\value{
an \code{sf} object
}
\description{
Download a zipped \emph{shapefile} and returns it as an \code{sf} object.
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
