% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftp.R
\name{gds_ftp}
\alias{gds_ftp}
\title{ftp directory for GDS accession}
\usage{
gds_ftp(accession)
}
\arguments{
\item{accession}{character.}
}
\value{
character. path to the ftp directory, without trailing \code{/}
}
\examples{
\\donotrun{
    gds_ftp('GDS479') \%>\% browseURL()
}

}
