% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geo_info.R
\name{get_geo_info}
\alias{get_geo_info}
\title{Get geographic information for coordinates}
\usage{
get_geo_info(.Data = NULL, .latitude = NULL, .longitude = NULL,
  .countries = list())
}
\arguments{
\item{.Data}{A data.table.
You can either specify an input data.table with columns named \code{longitude}
and \code{latitude} or pass on two equal-length vectors to \code{.longitude} and \code{.latitude}.}

\item{.latitude}{Numeric vector of latitudes.}

\item{.longitude}{Numeric vector of longitudes.}

\item{.countries}{A list with geographic data from \code{\link[raster:getData]{raster::getData()}}.}

\item{.sites}{SpatialPolygons object.
Spatial object to test if points lie inside sites.}
}
\value{
A data.table with new columns for the geographic informations.
}
\description{
\code{get_geo_info} returns geographical information like country, tk25
from coordinates.
}
\examples{
geographic_data_de <- raster::getData("GADM", country = "DE", level = 2)
get_geo_info(.longitude = 11, .latitude = 48, .countries = list(geographic_data_de))
file.remove("GADM_2.8_DEU_adm2.rds")
}
