% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geographic_data.R
\name{get_geographic_data}
\alias{get_geographic_data}
\title{Get geographic information for coordinates}
\usage{
get_geographic_data(
  .Data = NULL,
  .latitude = NULL,
  .longitude = NULL,
  .countries = list()
)
}
\arguments{
\item{.Data}{A data.table.
You can either specify an input data.table with columns named \code{longitude}
and \code{latitude} or pass on two equal-length vectors to \code{.longitude} and \code{.latitude}.}

\item{.latitude}{Numeric vector of latitudes.}

\item{.longitude}{Numeric vector of longitudes.}

\item{.countries}{A list with geographic data from \code{\link[raster:getData]{raster::getData()}}.}
}
\value{
A list containing named entries NAME_0, NAME_1, NAME_2, NAME_3, NAME_4.
See \code{\link[raster:getData]{raster::getData()}}.
}
\description{
Get geographic information for coordinates
}
\examples{
library(data.table)
Data <- data.table(latitude = 48:50, longitude = 11:13)
geographic_data_de <- raster::getData("GADM", country = "DE", level = 2)
get_geographic_data(Data, .countries = list(geographic_data_de))
file.remove("GADM_2.8_DEU_adm2.rds")
}
\seealso{
\code{\link[raster:getData]{raster::getData()}}
}
