% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-gse.R
\name{parse_gse_soft}
\alias{parse_gse_soft}
\title{read GSE soft file (contains platform annotation)}
\usage{
parse_gse_soft(soft_file, verbose = T)
}
\arguments{
\item{path}{string. path to file}
}
\value{
list
\enumerate{
\item accession: string. platform accession
\item info: tibble. chip column name information
\item table: tibble. chip annotation
}
}
\details{
for now, we drop probes which haven't been mapped to a symbool (mapping multiple symbols is okay)
}
\examples{

}
\seealso{
\link{read_gse_soft}
}
