% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-gse.R
\name{read_gse_matrix}
\alias{read_gse_matrix}
\title{read GSE series matrix file (contains expression data)}
\usage{
read_gse_matrix(matrix_file, verbose = F)
}
\arguments{
\item{path}{string. path to file}
}
\value{
tibble. the first variable must be \code{ID_REF}, others are gene expression value of each sample
}
\details{
for now, we don't collect sample meta data
}
\examples{

}
\seealso{
\link{write_txt}
}
